# backend/docs/API_DOCUMENTATION.yaml
openapi: 3.0.3
info:
  title: CozyVTT API
  description: |
    RESTful API for CozyVTT — a self-hosted, browser-based Virtual Tabletop platform.

    ## Authentication
    All endpoints (except `/health`, `/api/setup/*`, and public auth routes) require authentication via session cookies.
    Login via `/api/auth/login` to obtain a session cookie.

    ## MFA Flow
    If a user has MFA enabled, `/api/auth/login` returns `{ mfaRequired: true }` instead of a full session.
    The client must then call `/api/auth/mfa/verify-login` with a TOTP code or backup code to complete authentication.

    ## Authorization
    - **Platform roles**: `ADMIN` (full system access), `USER` (campaign access)
    - **Campaign roles**: `DM` (full campaign control), `PLAYER` (limited interaction), `SPECTATOR` (read-only)

    ## Rate Limiting
    - Global API limit: 300 requests per minute per IP (all `/api/*` routes)
    - Authentication endpoints: 5 requests per 15 minutes
    - MFA setup verification: 5 attempts per 15 minutes
    - MFA login verification: 5 attempts per 15 minutes (separate limit)
    - Asset uploads: 30 requests per minute per user (configurable via `ASSET_UPLOAD_RATE_LIMIT`)
    - Chat messages: 5 per minute (WebSocket)
    - Dice rolls: 30 per minute (WebSocket)
    - Token movement: 60 per second (WebSocket, throttled)

    Rate-limited responses return HTTP 429 with `RateLimit-*` standard headers.

    ## Forced password changes
...

    ## Real-time Events
...

  version: 1.2.1
...
servers:
...
security:
  - cookieAuth: []
...
paths:
...
  /api/campaigns/{campaignId}/creatures:
    get:
      tags: [Creatures]
      summary: List creature templates
      description: |
        Returns SRD (global) + campaign-specific custom creature templates.
        Supports search, source filter, challenge rating filter, and pagination.
        Any campaign member can list.
      operationId: listCreatures
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CampaignId'
        - name: search
          in: query
          schema:
            type: string
          description: Filter by name (case-insensitive partial match)
        - name: source
          in: query
          schema:
            type: string
            enum: [srd, custom]
          description: Filter by source
        - name: cr
          in: query
          schema:
            type: string
          description: Filter by challenge rating
        - name: gameSystem
          in: query
          schema:
            type: string
          description: Filter by game system
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Paginated creature list
          content:
            application/json:
              schema:
                type: object
                properties:
                  creatures:
                    type: array
                    items:
                      $ref: '#/components/schemas/CreatureTemplate'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Creatures]
      summary: Create custom creature
      description: Create a new campaign-specific custom creature template (DM only).
      operationId: createCreature
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CampaignId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - statBlock
              properties:
                name:
                  type: string
                statBlock:
                  type: object
                gameSystem:
                  type: string
                challengeRating:
                  type: string
                creatureType:
                  type: string
                alignment:
                  type: string
                imageUrl:
                  type: string
                size:
                  type: object
                  properties:
                    width:
                      type: integer
                    height:
                      type: integer
                disposition:
                  type: string
                  enum: [hostile, friendly, neutral]
                displayMode:
                  type: string
                  enum: [pog, top-down, full-art]
      responses:
        '201':
          description: Creature template created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatureTemplate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/campaigns/{campaignId}/creatures/{creatureId}:
    get:
      tags: [Creatures]
      summary: Get creature template
      description: Get a single creature template by ID. Any campaign member can view.
      operationId: getCreature
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CampaignId'
        - name: creatureId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Creature template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatureTemplate'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Creatures]
      summary: Update custom creature
      description: Update a campaign-specific custom creature template (DM only). Cannot edit SRD creatures (returns 403).
      operationId: updateCreature
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CampaignId'
        - name: creatureId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                statBlock:
                  type: object
                challengeRating:
                  type: string
                creatureType:
                  type: string
                alignment:
                  type: string
                imageUrl:
                  type: string
                size:
                  type: object
                disposition:
                  type: string
                displayMode:
                  type: string
      responses:
        '200':
          description: Creature template updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatureTemplate'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Creatures]
      summary: Delete custom creature
      description: Delete a campaign-specific custom creature template (DM only). Cannot delete SRD creatures (returns 403).
      operationId: deleteCreature
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CampaignId'
        - name: creatureId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Creature deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/campaigns/{campaignId}/creatures/{creatureId}/duplicate:
    post:
      tags: [Creatures]
      summary: Duplicate creature
      description: Duplicate any creature (including SRD) as a new custom creature in this campaign (DM only).
      operationId: duplicateCreature
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CampaignId'
        - name: creatureId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Duplicated creature template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatureTemplate'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/campaigns/{campaignId}/creatures/favorites/list:
    get:
      tags: [Creatures]
      summary: List creature favorites
      description: Returns creature IDs and full creature objects that the current user has favorited in this campaign.
      operationId: listCreatureFavorites
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CampaignId'
      responses:
        '200':
          description: Favorite creature list
          content:
            application/json:
              schema:
                type: object
                properties:
                  favoriteIds:
                    type: array
                    items:
                      type: string
                      format: uuid
                  creatures:
                    type: array
                    items:
                      $ref: '#/components/schemas/CreatureTemplate'
  /api/campaigns/{campaignId}/creatures/{creatureId}/favorite:
    post:
      tags: [Creatures]
      summary: Toggle creature favorite
      description: Toggle favorite for the current user in this campaign. If already favorited, removes it; otherwise adds it.
      operationId: toggleCreatureFavorite
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CampaignId'
        - name: creatureId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Favorite toggled
          content:
            application/json:
              schema:
                type: object
                properties:
                  favorited:
                    type: boolean
  /api/campaigns/{campaignId}/creatures/seed/status:
    get:
      tags: [Creatures]
      summary: Check SRD seed status
      description: Returns whether SRD creatures have been seeded. Any campaign member can check.
      operationId: getSrdSeedStatus
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CampaignId'
      responses:
        '200':
          description: Seed status
          content:
            application/json:
              schema:
                type: object
                properties:
                  seeded:
                    type: boolean
                  count:
                    type: integer
                  seedInProgress:
                    type: boolean
  /api/campaigns/{campaignId}/creatures/seed:
    post:
      tags: [Creatures]
      summary: Seed SRD creatures
      description: |
        Fetch SRD creatures from Open5e and populate the library (DM only).
        Safe to call multiple times — existing SRD creatures are not duplicated.

        Existing SRD creatures stored before hit points were tracked are updated
        in place with `hpMax` / `hitDice`; only those missing keys are added, and
        custom creatures are never modified.

        Returns 409 if seeding is already in progress.
      operationId: seedSrdCreatures
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CampaignId'
      responses:
        '200':
          description: Seeding complete
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  fetched:
                    type: integer
                    description: Monsters returned by Open5e
                  created:
                    type: integer
                    description: New SRD creatures inserted
                  updated:
                    type: integer
                    description: Existing SRD creatures backfilled with hit points
                  skipped:
                    type: integer
                    description: Existing SRD creatures left unchanged
                  alreadyExisted:
                    type: integer
        '409':
          description: Seeding already in progress

  # ============================================
  # TOKEN TEMPLATE ENDPOINTS
  # ============================================
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: cozyvtt.sid
      description: Session cookie obtained via /api/auth/login
  parameters:
    CampaignId:
      name: campaignId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Campaign UUID
...
  schemas:
...
    CreatureTemplate:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        gameSystem:
          type: string
          nullable: true
        source:
          type: string
          enum: [srd, custom]
          description: "'srd' = imported from Open5e (global, read-only), 'custom' = campaign-specific"
        challengeRating:
          type: string
          nullable: true
        creatureType:
          type: string
          nullable: true
        alignment:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        statBlock:
          type: object
          description: |
            Game-system-agnostic stat block. Common keys: `ac`, `hpMax`, `hitDice`,
            `speed`, `abilities`, `traits`, `actions`. `hpMax` seeds the hit points
            of tokens placed from this creature (defaults to 10 when absent, as on
            stat blocks saved before hit points were tracked).
        size:
          type: object
          properties:
            width:
              type: integer
              default: 1
            height:
              type: integer
              default: 1
        disposition:
          type: string
          enum: [hostile, friendly, neutral]
          default: hostile
        displayMode:
          type: string
          enum: [pog, top-down, full-art]
          default: pog
        campaignId:
          type: string
          format: uuid
          nullable: true
          description: "null for SRD creatures (global), set for custom creatures"
        createdById:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: BadRequest
          description: Short machine-readable error identifier
        message:
          type: string
          example: One or more fields failed validation.
          description: Human-readable error description
        details:
          type: object
          nullable: true
          additionalProperties: true
          description: Optional structured details (e.g. per-field Zod issues)
...
  responses:
    BadRequest:
      description: Bad Request - Invalid input or validation failure.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: BadRequest
            message: Required field "name" is missing.
...
    Forbidden:
      description: Forbidden - Authenticated but insufficient permissions.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Forbidden
            message: You do not have permission to perform this action.
...
    NotFound:
      description: Not Found - Resource does not exist or you cannot see it.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: NotFound
            message: Campaign not found.
...